Move existing renderer to rayTracer directory
