+++ src/routes/chapters/01-common-concepts/+page.svelte
@@ -2,7 +2,7 @@
   import { onMount, getContext } from 'svelte';
 
   import { Green } from '$lib/colors';
-  import { Renderer } from '$lib/renderer';
+  import { Renderer } from '$lib/rayTracer/renderer';
 
   const context = getContext<{ getCanvas: () => HTMLCanvasElement }>('canvas');
   console.log(context);
